requirements.txt
Flask==2.3.3
requests==2.31.0
Pillow==10.0.1
Werkzeug==2.3.7
vosk==0.3.45
webrtcvad==2.0.10
pydub==0.25.1
ffmpeg-python==0.2.0
speechrecognition==3.10.0

